requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
folium>=0.12.0
geopandas>=0.10.0
plotly>=5.5.0
requests>=2.26.0
beautifulsoup4>=4.9.0
selenium>=4.0.0
webdriver-manager>=3.5.0
python-dotenv>=0.19.0
jupyter>=1.0.0
flask>=2.0.0
flask-sqlalchemy>=2.5.0
flask-wtf>=1.0.0
flask-login>=0.5.0 